13/11/2023 import teaching

+++ Meet-My-Lecture/src/main/java/com/springboot/meetMyLecturer/modelDTO/EmptySlotDTO.java
@@ -7,7 +7,7 @@ import java.sql.Date;
 
 @Data
 public class EmptySlotDTO {
-    private int emptySlotId;
+    private Long emptySlotId;
 
     private Date dateStart;
